Refactor package structure for Jackson tests
